+++ test/number/number_system_test.exs
@@ -23,4 +23,22 @@ defmodule Number.System.Test do
     assert is_list(likes)
     assert Enum.count(likes) > 100
   end
+
+  test "that locale u extension number system overrides default" do
+    assert TestBackend.Cldr.Number.to_string(1234, locale: "th-u-nu-thai") == {:ok, "๑,๒๓๔"}
+    assert TestBackend.Cldr.Number.to_string(1234, locale: "th-u-nu-latn") == {:ok, "1,234"}
+  end
+
+  test "that number_system parameter overrides the locale u number system" do
+    assert TestBackend.Cldr.Number.to_string(1234, locale: "th-u-nu-latn", number_system: :thai) ==
+      {:ok, "๑,๒๓๔"}
+  end
+
+  test "that a locale u number system that is not valid for a locale returns an error" do
+    assert TestBackend.Cldr.Number.to_string(1234, locale: "en-AU-u-nu-thai") ==
+    {:error,
+     {Cldr.UnknownNumberSystemError,
+      "The number system :thai is unknown for the locale named \"en-AU\". Valid number systems are %{default: :latn, native: :latn}"}}
+  end
+
 end